pyproject.toml
[project]
name = "template-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pdfplumber>=0.11.9",
    "pydantic[email,timezone]>=2.12.5",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.13",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
